Added lyrics lookup for songs

+++ js/music.js
@@ -1,5 +1,22 @@
-const ytinfo = require('ytdl-getinfo');
 const ytdl = require('youtube-dl');
+
+
+const geniusId = "P16rUAwFfDqa6orLH8sd5_VGpWTDVNnxXwjkVdEGuQODF6_D2miuEPQPhmSXsik7";
+const geniusSecret = "5a9Kbhd31hyoQTN0Tr7cTds9PkrjZFXM5FDDrPZr8KFmBQn--kGS9iX3UOSVoMlHWcX0ZpFIqAWQpadtUd4t2g";
+const geniusClient = "l5zrX9XIDrJuz-kS1u7zS5sE81KzrH3qxZL5tAvprE9GG-L1KYlZklQDXL6wf3sn";
+const credentials = {
+  client: {id: geniusId, secret: geniusSecret},
+  auth: {tokenHost: "https://api.genius.com"}
+};
+const geniusRequest = {
+  hostname: 'api.genius.com',
+  path: '/search/',
+  headers:
+      {"Accept": "application/json", "Authorization": "Bearer " + geniusClient},
+  method: "GET"
+};
+// const oauth2 = require('simple-oauth2').create(credentials);
+const https = require('https');
 
 var initialized = false;
 
@@ -9,6 +26,29 @@ var prefix, Discord, client, command, common;
 var myPrefix, helpMessage;
 
 var broadcasts = {};
+
+const special = {
+  'vi rap': {
+    cmd: 'vi',
+    url: "https://www.youtube.com/watch?v=c1NoTNCiomU",
+    file: '/home/discord/SpikeyBot-Discord/js/sounds/viRap.webm'
+  },
+  'airhorn': {
+    cmd: 'airhorn',
+    url: "",
+    file: '/home/discord/SpikeyBot-Discord/js/sounds/airhorn.mp3'
+  },
+  'rickroll': {
+    cmd: 'rickroll',
+    url: "https://www.youtube.com/watch?v=dQw4w9WgXcQ",
+    file: '/home/discord/SpikeyBot-Discord/js/sounds/rickRoll.webm'
+  },
+  'kokomo': {
+    cmd: 'kokomo',
+    url: "https://www.youtube.com/watch?v=fJWmbLS2_ec",
+    file: "/home/discord/SpikeyBot-Discord/js/sounds/kokomo.webm"
+  }
+};
 
 const ytdlOpts =
     ['-f bestaudio/best', '--no-playlist', '--default-search=auto'];
@@ -27,6 +67,7 @@ exports.begin = function(prefix_, Discord_, client_, command_, common_) {
   command.on('skip', commandSkip, true);
   command.on(['queue', 'playing'], commandQueue, true);
   command.on(['remove', 'dequeue'], commandRemove, true);
+  command.on('lyrics', commandLyrics);
 
   command.on('kokomo', msg => {
     msg.content = "?play kokomo";
@@ -34,6 +75,14 @@ exports.begin = function(prefix_, Discord_, client_, command_, common_) {
   });
   command.on('vi', msg => {
     msg.content = "?play vi rap";
+    command.trigger('play', msg);
+  });
+  command.on('airhorn', msg => {
+    msg.content = "?play airhorn";
+    command.trigger('play', msg);
+  });
+  command.on('rickroll', msg => {
+    msg.content = "?play rickroll";
     command.trigger('play', msg);
   });
 
@@ -91,7 +140,7 @@ function formNum(num) {
   return tmpString.reverse().join('');
 }
 // Add a song to the given broadcast's queue and start playing it not already.
-function queueSong(broadcast, song, msg, info) {
+function enqueueSong(broadcast, song, msg, info) {
   broadcast.queue.push({request: msg, song: song, info: info});
   if (broadcast.voice) {
     try {
@@ -107,7 +156,7 @@ function queueSong(broadcast, song, msg, info) {
           broadcast.voice = conn;
           try {
             startPlaying(broadcast);
-          } catch(err) {
+          } catch (err) {
             console.log(err);
             reply(msg, "Failed to start music stream!");
             command.trigger('stop', msg);
@@ -126,21 +175,14 @@ function startPlaying(broadcast) {
   }
   if (broadcast.queue.length === 0) {
     command.trigger('stop', broadcast.current.request);
-    broadcast.current.request.channel.send(
-        "Queue is empty!\n```\nSee you later!\n```");
+    broadcast.current.request.channel.send("`Queue is empty!`");
     return;
   }
   broadcast.isLoading = true;
   broadcast.skips = {};
   broadcast.current = broadcast.queue.splice(0, 1)[0];
   try {
-    if (broadcast.current.info) {
-      broadcast.current.stream = ytdl(broadcast.current.info.url, ytdlOpts);
-    } else {
-      broadcast.current.stream = ytdl(broadcast.current.song, ytdlOpts);
-    }
-    broadcast.broadcast.playStream(broadcast.current.stream)
-        .on('end', function() { endSong(broadcast); });
+    makeBroadcast(broadcast);
     broadcast.voice.playBroadcast(broadcast.broadcast);
   } catch (err) {
     console.log(err);
@@ -151,20 +193,64 @@ function startPlaying(broadcast) {
 
   if (typeof broadcast.current.info !== 'undefined') {
     var embed = formatSongInfo(broadcast.current.info);
-    embed.setTitle("Now playing [" + broadcast.queue.length + " left in queue]");
+    embed.setTitle(
+        "Now playing [" + broadcast.queue.length + " left in queue]");
     broadcast.current.request.channel.send(embed);
     broadcast.current.stream.on(
         'info', info => { broadcast.isLoading = false; });
   } else {
-    // reply(broadcast.current.request, "Playing next song");
-    broadcast.current.stream.on('info', info => {
-      broadcast.isLoading = false;
-      broadcast.current.info = info;
-      var embed = formatSongInfo(broadcast.current.info);
-      embed.setTitle(
-          "Now playing [" + broadcast.queue.length + " left in queue]");
-      broadcast.current.request.channel.send(embed);
-    });
+    if (special[broadcast.current.song]) {
+      if (!special[broadcast.current.song].url) {
+        broadcast.isLoading = false;
+        var embed = new Discord.RichEmbed();
+        embed.setTitle(
+            "Now playing [" + broadcast.queue.length + " left in queue]");
+        embed.setColor([50, 200, 255]);
+        embed.setDescription(broadcast.current.song);
+        broadcast.current.request.channel.send(embed);
+      } else {
+      ytdl.getInfo(
+          special[broadcast.current.song].url, ytdlOpts, (err, info) => {
+            if (err) {
+              console.log(err);
+              reply(
+                  msg,
+                  "Oops, something went wrong while getting info for this song!");
+            } else {
+              broadcast.isLoading = false;
+              broadcast.current.info = info;
+              var embed = formatSongInfo(broadcast.current.info);
+              embed.setTitle(
+                  "Now playing [" + broadcast.queue.length + " left in queue]");
+              broadcast.current.request.channel.send(embed);
+            }
+          });
+      }
+    } else {
+      broadcast.current.stream.on('info', info => {
+        broadcast.isLoading = false;
+        broadcast.current.info = info;
+        var embed = formatSongInfo(broadcast.current.info);
+        embed.setTitle(
+            "Now playing [" + broadcast.queue.length + " left in queue]");
+        broadcast.current.request.channel.send(embed);
+      });
+    }
+  }
+}
+function makeBroadcast(broadcast) {
+  if (special[broadcast.current.song]) {
+    broadcast.broadcast.playFile(special[broadcast.current.song].file)
+        .on('end', function() { endSong(broadcast); });
+  } else {
+    if (broadcast.current.info) {
+      broadcast.current.stream = ytdl(broadcast.current.info.url, ytdlOpts);
+    } else {
+      broadcast.current.stream = ytdl(broadcast.current.song, ytdlOpts);
+    }
+
+    broadcast.broadcast.playStream(broadcast.current.stream)
+        .on('end', function() { endSong(broadcast); });
   }
 }
 // Triggered when a song has finished playing.
@@ -190,37 +276,50 @@ function commandPlay(msg) {
     } else {
       song = song.replace(' ', '');
     }
-    var loadingMsg;
-    reply(msg, "Loading " + song + "\nPlease wait...")
-        .then(msg => loadingMsg = msg);
     if (!broadcasts[msg.guild.id]) {
+      reply(msg, "Loading " + song + "\nPlease wait...")
+          .then(msg => msg.delete(10000));
       broadcasts[msg.guild.id] = {
         queue: [],
         skips: {},
         isPlaying: false,
         broadcast: client.createVoiceBroadcast()
       };
-      queueSong(broadcasts[msg.guild.id], song, msg);
+      enqueueSong(broadcasts[msg.guild.id], song, msg);
     } else {
-      ytdl.getInfo(song, ytdlOpts, (err, info) => {
-        if (err) {
-          reply(
-              msg, "Oops, something went wrong while searching for that song!");
-          console.log(err);
-        } else if (info._duration_raw == 0) {
-          reply(msg, "Sorry, but I can't play live streams currently.");
-        } else {
-          if (broadcasts[msg.guild.id].isPlaying) {
-            var embed = formatSongInfo(info);
-            embed.setTitle(
-                "Enqueuing " + song + " [" +
-                (broadcasts[msg.guild.id].queue.length + 1) + " in queue]");
-            msg.channel.send(mention(msg), embed);
-            queueSong(broadcasts[msg.guild.id], song, msg, info);
+      if (special[song]) {
+        var embed = new Discord.RichEmbed();
+        embed.setTitle(
+            "Enqueuing " + song + " [" +
+            (broadcasts[msg.guild.id].queue.length + 1) + " in queue]");
+        embed.setColor([50, 200, 255]);
+        msg.channel.send(mention(msg), embed);
+        enqueueSong(broadcasts[msg.guild.id], song, msg);
+      } else {
+        var loadingMsg;
+        reply(msg, "Loading " + song + "\nPlease wait...")
+            .then(msg => loadingMsg = msg);
+        ytdl.getInfo(song, ytdlOpts, (err, info) => {
+          if (err) {
+            reply(
+                msg,
+                "Oops, something went wrong while searching for that song!");
+            console.log(err);
+          } else if (info._duration_raw == 0) {
+            reply(msg, "Sorry, but I can't play live streams currently.");
+          } else {
+            if (broadcasts[msg.guild.id].isPlaying) {
+              var embed = formatSongInfo(info);
+              embed.setTitle(
+                  "Enqueuing " + song + " [" +
+                  (broadcasts[msg.guild.id].queue.length + 1) + " in queue]");
+              msg.channel.send(mention(msg), embed);
+              enqueueSong(broadcasts[msg.guild.id], song, msg, info);
+            }
           }
-        }
-        if (loadingMsg) loadingMsg.delete();
-      });
+          if (loadingMsg) loadingMsg.delete();
+        });
+      }
     }
   }
 }
@@ -257,18 +356,30 @@ function commandQueue(msg) {
   } else {
     var emebed;
     if (broadcasts[msg.guild.id].current) {
-      embed = formatSongInfo(broadcasts[msg.guild.id].current.info);
+      if (broadcasts[msg.guild.id].current.info) {
+        embed = formatSongInfo(broadcasts[msg.guild.id].current.info);
+      } else {
+        embed = new Discord.RichEmbed();
+        embed.setColor([50, 200, 255]);
+        embed.setDescription(broadcasts[msg.guild.id].current.song);
+      }
       embed.setTitle("Current Song Queue");
     } else {
       embed = new Discord.RichEmbed();
     }
-    embed.addField(
-        "Queue", broadcasts[msg.guild.id]
-                     .queue
-                     .map(function(obj, index) {
-                       return (index + 1) + ") " + obj.info.title;
-                     })
-                     .join('\n'));
+    if (broadcasts[msg.guild.id].queue.length > 0) {
+      embed.addField(
+          "Queue", broadcasts[msg.guild.id]
+                       .queue
+                       .map(function(obj, index) {
+                         if (obj.info) {
+                           return (index + 1) + ") " + obj.info.title;
+                         } else {
+                           return (index + 1) + ") " + obj.song;
+                         }
+                       })
+                       .join('\n'));
+    }
     msg.channel.send(embed);
   }
 }
@@ -296,5 +407,132 @@ function commandRemove(msg) {
         reply(msg, "Dequeued #" + index + ": " + removed.info.title);
       }
     }
+  }
+}
+function commandLyrics(msg) {
+  var song = msg.content.replace(myPrefix + "lyrics", '');
+  if (song.length <= 1) {
+    reply(msg, "Please specify a song.");
+    return;
+  }
+  song = song.replace(' ', '');
+  var thisReq = geniusRequest;
+  thisReq.path = "/search?q=" + encodeURIComponent(song);
+  var req = https.request(thisReq, function(response) {
+    var content = '';
+    response.on('data', function(chunk) { content += chunk; });
+    response.on('end', function() {
+      if (response.statusCode == 200) {
+        msg.channel.send("Search successful").then(msg => {
+          msg.delete(2000);
+        });
+        var parsed = JSON.parse(content);
+        if (parsed.response.hits.length === 0) {
+          reply(msg, "Failed to find lyrics. No matches found.");
+        } else {
+          reqLyricsURL(msg, parsed.response.hits[0].result.id);
+        }
+      } else {
+        msg.channel.send(
+            response.statusCode + "```json\n" +
+            JSON.stringify(response.headers, null, 2) + "```\n```html\n" +
+            content + "\n```");
+      }
+    });
+    response.on('close', function() {
+      common.LOG("Genius request closed! " + content.length, "Music");
+    });
+    response.on('error', function() {
+      common.LOG("Genius request errored! " + content.length, "Music");
+    });
+  });
+  req.end();
+  req.on('error', function(e) { common.ERROR(e, "Music"); });
+}
+function reqLyricsURL(msg, id) {
+  var thisReq = geniusRequest;
+  thisReq.path = "/songs/" + id + "?text_format=plain";
+  var req = https.request(thisReq, function(response) {
+    var content = '';
+    response.on('data', function(chunk) { content += chunk; });
+    response.on('end', function() {
+      if (response.statusCode == 200) {
+        msg.channel.send("Lyrics url fetched.").then(msg => {
+          msg.delete(2000);
+        });
+        var parsed = JSON.parse(content);
+        fetchLyricsPage(
+            msg, parsed.response.song.url, parsed.response.song.full_title,
+            parsed.response.song.song_art_image_thumbnail_url);
+      } else {
+        msg.channel.send(
+            response.statusCode + "```json\n" +
+            JSON.stringify(response.headers, null, 2) + "```\n```html\n" +
+            content + "\n```");
+      }
+    });
+    response.on('close', function() {
+      common.LOG("Genius request closed! " + content.length, "Music");
+    });
+    response.on('error', function() {
+      common.LOG("Genius request errored! " + content.length, "Music");
+    });
+  });
+  req.end();
+  req.on('error', function(e) { common.ERROR(e, "Music"); });
+}
+function fetchLyricsPage(msg, url, title, thumb) {
+  var URL = url.match(/https:\/\/([^\/]*)(.*)/);
+  const thisReq = {hostname: URL[1], path: URL[2], method: "GET"};
+  var req = https.request(thisReq, function(response) {
+    var content = '';
+    response.on('data', function(chunk) { content += chunk; });
+    response.on('end', function() {
+      if (response.statusCode == 200) {
+        msg.channel.send("Lyrics page fetched.").then(msg => {
+          msg.delete(2000);
+        });
+        stripLyrics(msg, content, title, url, thumb);
+      } else {
+        msg.channel.send(
+            response.statusCode + "```json\n" +
+            JSON.stringify(response.headers, null, 2) + "```\n```html\n" +
+            content + "\n```");
+      }
+    });
+    response.on('close', function() {
+      common.LOG("Genius request closed! " + content.length, "Music");
+    });
+    response.on('error', function() {
+      common.LOG("Genius request errored! " + content.length, "Music");
+    });
+  });
+  req.end();
+  req.on('error', function(e) { common.ERROR(e, "Music"); });
+}
+function stripLyrics(msg, content, title, url, thumb) {
+  try {
+    var body = content.match(/<!--sse-->([\s\S]*?)<!--\/sse-->/gm)[1];
+    var lyrics = body.match(/^([^<]*)<|>([^<]*)<|>([^<]*)$/g)
+                     .slice(1)
+                     .join('')
+                     .replace(/<>|^\s*|^>\s*/gm, '');
+    var embed = new Discord.RichEmbed();
+    if (title) embed.setTitle(title);
+    if (url) {
+      embed.setFooter(url);
+      embed.setURL(url);
+    }
+    if (thumb) {
+      embed.setThumbnail(thumb);
+    }
+    for (var i = 0; i < 25 && lyrics.length > i * 1024; i++) {
+      embed.addField('\u200B', lyrics.substr(i * 1024, 1024), true);
+    }
+    embed.setColor([0, 255, 255]);
+    msg.channel.send(embed);
+  } catch(err) {
+    console.log(err);
+    msg.channel.send("FAILED to parse lyrics: " + err.message);
   }
 }
